feat(account): move DTO conversion to UseCase from AccountService.getAccountViewDto
