+++ src/main/kotlin/com/example/workflow/feature/account/usecase/GetAccountUseCase.kt
@@ -1,5 +1,7 @@
 package com.example.workflow.feature.account.usecase
 
+import com.example.workflow.core.account.Account
+import com.example.workflow.core.account.toViewDto
 import com.example.workflow.feature.account.model.AccountViewDto
 import com.example.workflow.feature.account.service.AccountService
 import org.springframework.stereotype.Service
@@ -15,9 +17,9 @@ class GetAccountUseCase(
 
     fun execute(): Result {
         val accountId: UUID = accountService.getCurrentAccountId()
-        val accountViewDto: AccountViewDto = accountService.getAccountViewDto(accountId)
+        val account: Account = accountService.getAccount(accountId)
         return Result(
-            accountViewDto = accountViewDto
+            accountViewDto = account.toViewDto()
         )
     }
 }
