+++ src/test/kotlin/com/example/workflow/unit/feature/account/usecase/GetAccountUseCaseTest.kt
@@ -1,11 +1,15 @@
 package com.example.workflow.unit.feature.account.usecase
 
+import com.example.workflow.core.account.Account
+import com.example.workflow.core.account.toViewDto
 import com.example.workflow.feature.account.model.AccountViewDto
 import com.example.workflow.feature.account.service.AccountService
 import com.example.workflow.feature.account.usecase.GetAccountUseCase
 import com.example.workflow.support.annotation.UnitTest
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.mockkStatic
+import io.mockk.unmockkStatic
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
@@ -31,20 +35,32 @@ class GetAccountUseCaseTest {
 
     @Nested
     inner class ExecuteMethod {
+        @BeforeEach
+        fun setUp() {
+            mockkStatic(Account::toViewDto)
+        }
+
+        @AfterEach
+        fun tearDown() {
+            unmockkStatic(Account::toViewDto)
+        }
+
         @Test
         fun `execute method should return account`() {
             // Arrange
             val accountId: UUID = UUID.randomUUID()
-            val accountViewDto: AccountViewDto = mockk()
+            val accountViewDtoMock: AccountViewDto = mockk()
+            val accountMock: Account = mockk()
 
             every { accountServiceMock.getCurrentAccountId() } returns accountId
-            every { accountServiceMock.getAccountViewDto(accountId) } returns accountViewDto
+            every { accountServiceMock.getAccount(accountId) } returns accountMock
+            every { accountMock.toViewDto() } returns accountViewDtoMock
 
             // Act
             val actual: GetAccountUseCase.Result = getAccountUseCase.execute()
 
             // Assert
-            assertEquals(accountViewDto, actual.accountViewDto)
+            assertEquals(accountViewDtoMock, actual.accountViewDto)
         }
     }
 }
